Feature/RedesignSpaceRoads

+++ Ship_Game/Empire_Trade.cs
@@ -282,7 +282,7 @@ namespace Ship_Game
         // FB - Refit some idle freighters to better ones, if unlocked
         public void TriggerFreightersRefit()
         {
-            if (ManualTrade || TotalFreighters / FreighterCap <= 0.75f)
+            if (ManualTrade || TotalFreighters / (float)FreighterCap <= 0.75f)
                 return;
 
             IShipDesign betterFreighter = ShipBuilder.PickFreighter(this, FastVsBigFreighterRatio);
@@ -296,7 +296,7 @@ namespace Ship_Game
         // Percentage to check if there is better suited freighter model available
         public void CheckForRefitFreighter(Ship freighter, int percentage, IShipDesign betterFreighter = null)
         {
-            if (!ManualTrade && !RandomMath.RollDice(percentage) && TotalFreighters / FreighterCap > 0.5f)
+            if (!ManualTrade && !RandomMath.RollDice(percentage) && TotalFreighters / (float)FreighterCap > 0.5f)
             {
                 if (betterFreighter == null)
                     betterFreighter = ShipBuilder.PickFreighter(this, FastVsBigFreighterRatio);
